Business central advanced settings PR comments fix
